requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blockbuster==1.5.24
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cryptography==44.0.2
dataclasses-json==0.6.7
deepgram-sdk
filetype==1.2.0
forbiddenfruit==0.1.4
frozenlist==1.6.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
grpcio==1.72.0
grpcio-status==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.56
langchain-google-genai==2.1.3
langchain-text-splitters==0.3.8
langgraph==0.3.34
langgraph-api==0.1.16
langgraph-checkpoint==2.0.24
langgraph-cli==0.2.7
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.6
langgraph-sdk==0.1.63
langsmith==0.3.34
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.2.0
tavily-python==0.7.0
tenacity==9.1.2
tiktoken==0.9.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
streamlit>=1.36.0
flask>=2.0.0
pillow==10.4.0
gTTS==2.5.0
python-dotenv==1.1.0
SpeechRecognition==3.10.0
sounddevice==0.4.6
soundfile==0.12.1
scikit-learn==1.5.2
